requirements.txt
Django==1.7.6
argparse==1.2.1
coverage==3.7.1
django-crispy-forms==1.4.0
nose==1.3.4
wsgiref==0.1.2
